feat(esm): esm all the things
