+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@network-goods/hypercerts-sdk",
-  "version": "0.0.8",
+  "version": "0.0.9",
   "description": "SDK for hypercerts protocol",
   "repository": "git@github.com:Network-Goods/hypercerts-sdk.git",
   "author": "Network Goods",
@@ -59,7 +59,7 @@
     "shx": "^0.3.4",
     "ts-mocha": "^10.0.0",
     "ts-node": "^10.9.1",
-    "typescript": "^5.0.0-dev.20230109"
+    "typescript": "^4.9.0"
   },
   "resolutions": {
     "graphql": "^16.6.0"
@@ -74,18 +74,11 @@
     "graph:test": "echo \"Error: No tests\" && exit 1",
     "graph:watch": "graphql-codegen -w",
     "clean": "rm -rf ./lib",
-    "build": "npm run clean && npm run graph:build && npm run build:esm",
+    "build": "npm run clean && npm run graph:build:esm && npm run build:esm",
     "build:esm": "tsc",
     "prepack": "npm run build",
-    "test": "yarn jest",
+    "test": "NODE_OPTIONS=\"--experimental-vm-modules\" jest",
     "types:json": "yarn run json2ts -i 'resources/schema/' -o 'types' --cwd 'resources/schema' --no-declareExternallyReferenced"
   },
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./lib/src/index.d.ts",
-        "default": "./lib/src/index.js"
-      }
-    }
-  }
+  "type": "module"
 }
