+++ src/index.ts
@@ -1,13 +1,13 @@
 // Validation
-import { validateMetaData, validateClaimData } from "./validator";
-import { storeMetadata, storeData, getMetadata, getData, deleteMetadata } from "./operator";
+import { validateMetaData, validateClaimData } from "./validator/index.js";
+import { storeMetadata, storeData, getMetadata, getData, deleteMetadata } from "./operator/index.js";
 
 export { validateMetaData, validateClaimData, storeMetadata, storeData, getMetadata, getData, deleteMetadata };
 
 // Graph
-import { execute } from "../.graphclient";
-import { claimsByOwner, claimById, firstClaims } from "./claims";
-import { fractionsByOwner, fractionsByClaim } from "./fractions";
+import { execute } from "../.graphclient/index.js";
+import { claimsByOwner, claimById, firstClaims } from "./claims.js";
+import { fractionsByOwner, fractionsByClaim } from "./fractions.js";
 
 export { execute, claimsByOwner, claimById, firstClaims, fractionsByOwner, fractionsByClaim };
 
@@ -17,7 +17,7 @@ import { HyperCertMinterFactory, HypercertMinterABI } from "@network-goods/hyper
 export { HyperCertMinterFactory, HypercertMinterABI };
 
 // Types
-import type { HypercertClaimdata } from "../types/claimdata";
-import type { HypercertMetadata } from "../types/metadata";
+import type { HypercertClaimdata } from "../types/claimdata.mjs";
+import type { HypercertMetadata } from "../types/metadata.mjs";
 
 export type { HypercertMetadata, HypercertClaimdata };
