+++ src/validator/index.ts
@@ -2,10 +2,10 @@ import Ajv from "ajv";
 import metaData from "../../resources/schema/metadata.json";
 import claimData from "../../resources/schema/claimdata.json";
 
-import { HypercertMetadata } from "../../types/metadata";
-import { HypercertClaimdata } from "../../types/claimdata";
+import { HypercertMetadata } from "../../types/metadata.mjs";
+import { HypercertClaimdata } from "../../types/claimdata.mjs";
 
-const ajv = new Ajv(); // options can be passed, e.g. {allErrors: true}
+const ajv = new Ajv.default(); // options can be passed, e.g. {allErrors: true}
 ajv.addSchema(metaData, "metaData");
 ajv.addSchema(claimData, "claimData");
