+++ test/setup-tests.ts
@@ -1,0 +1,3 @@
+import dotenv from 'dotenv';
+
+dotenv.config({ path: './config.env.test' });
